Have TablesApiRequest Support Additional Query Parameters
